+++ lib/src/native_platform_interaction/faro_sdk_method_channel.dart
@@ -79,6 +79,11 @@ class MethodChannelFaroSdk extends FaroSdkPlatform {
   }
 
   @override
+  Future<void> purgeCrashReport() async {
+    await methodChannel.invokeMethod<void>('purgeCrashReport');
+  }
+
+  @override
   Future<Map<String, dynamic>?> getSessionRuntimeInfo({
     required bool claimSessionPersistence,
   }) {